ai-service/requirements.txt
groq
flask
flask-limiter
python-dotenv
bleach